feat: extract UI dialogs into separate classes for better maintainability
Extract dialog implementations from feature classes into dedicated dialog classes to improve code organization and separation of concerns. This refactoring makes the codebase more modular and easier to maintain by isolating UI logic from feature logic.

- Create dialog classes for all feature settings and selection dialogs
- Move dialog creation logic from feature classes to dialog constructors
- Maintain existing functionality while improving code structure
- Reduce coupling between feature logic and UI implementation
