+++ src/mindustrytool/features/chat/global/ChatFeature.java
@@ -5,15 +5,11 @@ import java.util.Optional;
 
 import arc.Core;
 import arc.Events;
-import arc.scene.event.Touchable;
 import arc.scene.ui.Dialog;
 import arc.util.Log;
 import arc.struct.Seq;
 import mindustry.Vars;
 import mindustry.gen.Icon;
-import arc.scene.ui.Label;
-import arc.scene.ui.Slider;
-import arc.scene.ui.layout.Table;
 import mindustry.ui.dialogs.BaseDialog;
 import mindustrytool.Utils;
 import mindustrytool.features.Feature;
@@ -103,111 +99,18 @@ public class ChatFeature implements Feature {
     @Override
     public Optional<Dialog> setting() {
         if (settingDialog == null) {
-            settingDialog = new BaseDialog("@chat.settings.title");
-            settingDialog.name = "chatSettingDialog";
-            settingDialog.addCloseButton();
-            settingDialog.closeOnBack();
-            settingDialog.shown(this::rebuildSettings);
+            settingDialog = new ChatSettingsDialog(this);
         }
         return Optional.of(settingDialog);
     }
 
-    private void rebuildSettings() {
-        Table cont = settingDialog.cont;
-        cont.clear();
-        cont.defaults().pad(6).left();
-        float width = Math.min(Core.graphics.getWidth() / 1.2f, 460f);
+    public ChatOverlay getOverlay() {
+        return overlay;
+    }
 
-        // Reset Position
-        cont.button("@chat.reset-overlay-position", () -> {
-            ChatConfig.x(0);
-            ChatConfig.y(0);
-
-            if (overlay != null) {
-                overlay.setPosition(0, 0);
-                overlay.keepInScreen();
-            }
-        }).size(240f, 50f).row();
-
-        // Opacity
-        Slider opacitySlider = new Slider(0.05f, 1f, 0.05f, false);
-        opacitySlider.setValue(ChatConfig.opacity());
-        Label opacityValue = new Label(String.format("%.0f%%", ChatConfig.opacity() * 100));
-
-        Table opacityContent = new Table();
-        opacityContent.touchable = Touchable.disabled;
-        opacityContent.add("@opacity").left().growX().padLeft(10).padRight(10);
-        opacityContent.add(opacityValue).padLeft(10f).right();
-
-        opacitySlider.changed(() -> {
-            ChatConfig.opacity(opacitySlider.getValue());
-            opacityValue.setText(String.format("%.0f%%", ChatConfig.opacity() * 100));
-            if (overlay != null)
-                overlay.rebuild();
-        });
-
-        cont.stack(opacitySlider, opacityContent).width(width).left().padTop(4f).row();
-
-        // Scale
-        Slider scaleSlider = new Slider(0.5f, 1.5f, 0.1f, false);
-        scaleSlider.setValue(ChatConfig.scale());
-        Label scaleValue = new Label(String.format("%.0f%%", ChatConfig.scale() * 100));
-
-        Table scaleContent = new Table();
-        scaleContent.touchable = Touchable.disabled;
-        scaleContent.add("@scale").left().growX().padLeft(10).padRight(10);
-        scaleContent.add(scaleValue).padLeft(10f).right();
-
-        scaleSlider.changed(() -> {
-            ChatConfig.scale(scaleSlider.getValue());
-            scaleValue.setText(String.format("%.0f%%", ChatConfig.scale() * 100));
-            if (overlay != null)
-                overlay.rebuild();
-        });
-
-        cont.stack(scaleSlider, scaleContent).width(width).left().padTop(4f).row();
-
-        // Width
-        Slider widthSlider = new Slider(0.1f, 1.0f, 0.1f, false);
-        widthSlider.setValue(ChatConfig.width());
-        Label widthValue = new Label(String.format("%.0f%%", ChatConfig.width() * 100));
-
-        Table widthContent = new Table();
-        widthContent.touchable = Touchable.disabled;
-        widthContent.add("@width").left().growX().padLeft(10).padRight(10);
-        widthContent.add(widthValue).padLeft(10f).right();
-
-        widthSlider.changed(() -> {
-            ChatConfig.width(widthSlider.getValue());
-            widthValue.setText(String.format("%.0f%%", ChatConfig.width() * 100));
-            if (overlay != null)
-                overlay.rebuild();
-        });
-
-        cont.stack(widthSlider, widthContent).width(width).left().padTop(4f).row();
-
-        // Height
-        Slider heightSlider = new Slider(0.1f, 1.0f, 0.1f, false);
-        heightSlider.setValue(ChatConfig.height());
-        Label heightValue = new Label(String.format("%.0f%%", ChatConfig.height() * 100));
-
-        Table heightContent = new Table();
-        heightContent.touchable = Touchable.disabled;
-        heightContent.add("@height").left().growX().padLeft(10).padRight(10);
-        heightContent.add(heightValue).padLeft(10f).right();
-
-        heightSlider.changed(() -> {
-            ChatConfig.height(heightSlider.getValue());
-            heightValue.setText(String.format("%.0f%%", ChatConfig.height() * 100));
-            if (overlay != null)
-                overlay.rebuild();
-        });
-
-        cont.stack(heightSlider, heightContent).width(width).left().padTop(4f).row();
-
-        // status toggle
-        cont.check("@chat.status", ChatConfig.status(), (v) -> {
-            ChatConfig.status(v);
-        }).row();
+    public void rebuildOverlay() {
+        if (overlay != null) {
+            overlay.rebuild();
+        }
     }
 }
